+++ components/contact.tsx
@@ -1,9 +1,20 @@
 "use client";
 
-import React from 'react'
-import SectionHeading from './shared/sectionHeading';
 import { useSectionInView } from '@/hooks/useSectionInView';
-import { FaPaperPlane } from 'react-icons/fa';
+import StylishButton from './shared/StylishButton';
+import SectionHeading from './shared/sectionHeading';
+
+import { Button } from "@/components/ui/button";
+import {
+    Sheet,
+    SheetClose,
+    SheetContent,
+    SheetDescription,
+    SheetFooter,
+    SheetHeader,
+    SheetTitle,
+    SheetTrigger,
+} from "@/components/ui/sheet";
 
 const ContactComponent = () => {
     const { ref } = useSectionInView("Contact");
@@ -11,15 +22,30 @@ const ContactComponent = () => {
     return (
         <section ref={ref} id='contact' className='lg:max-w-[50rem] xxs:w-64 xs:max-w-sm sm:max-w-sm sm:mb-0 md:mb-28 scroll-mt-28 mt-4 xxs:mt-10'>
             <SectionHeading heading={'Contact Me'} />
-            <div>
-                You can contact me directly by clicking over right side or via this form.
-            </div>
-            <form className='mt-10 flex flex-col justify-center items-center gap-2'>
-                <input type="email" className='h-14 w-full rounded-lg border-black/[0.1] shadow-md text-left my-auto pl-4 border-t-4 border-t-[#00d4ff]' placeholder='Your Email Address'/>
-                <input type="Subject" className='h-14 w-full rounded-lg border-black/[0.1] shadow-md text-left my-auto pl-4 border-t-4 border-t-[#00d4ff]' placeholder='Subject' />
-                <textarea className='h-52 min-h-min rounded-lg border-black/[0.1] my-3 p-4 w-full shadow-md border-t-4 border-t-[#00d4ff]' placeholder='Message to me' />
-                <button type='submit' className='h-12 text-white rounded-full bg-gray-950 outline-none transition-all w-32 gap-2 flex flex-row items-center justify-center'>Submit <FaPaperPlane /></button>
-            </form>
+            <p className='text-gray-800 text-lg'>Currently I&apos;m actively looking for any new opportunities, my inbox is always open. Whether you have a question or just want to say hi, I&apos;ll try my best to get back to you!</p>
+            <Sheet>
+                <SheetTrigger asChild>
+                    <span className='mt-10 flex flex-row justify-center items-center'>
+                    <StylishButton  />
+                    </span>
+                </SheetTrigger>
+                <SheetContent side={'right'}>
+                    <SheetHeader>
+                        <SheetTitle className='text-xl'>Hi There!</SheetTitle>
+                        <SheetDescription className='text-lg'>
+                           You can get in touch with me or follow me on these channels.
+                        </SheetDescription>
+                    </SheetHeader>
+                    <div className="grid gap-4 py-4">
+
+                    </div>
+                    <SheetFooter>
+                        <SheetClose asChild>
+                            <Button type="submit">Save changes</Button>
+                        </SheetClose>
+                    </SheetFooter>
+                </SheetContent>
+            </Sheet>
         </section>
     )
 }